Move component options to independent modules
