+++ packages/web-components/fast-components/src/component-definitions.ts
@@ -19,8 +19,8 @@ export * from "./skeleton/skeleton.definition";
 export * from "./slider/slider.definition";
 export * from "./slider-label/slider-label.definition";
 export * from "./switch/switch.definition";
-export * from "./tabs/tab-panel.definition";
-export * from "./tabs/tab.definition";
+export * from "./tab-panel/tab-panel.definition";
+export * from "./tab/tab.definition";
 export * from "./tabs/tabs.definition";
 export * from "./text-area/text-area.definition";
 export * from "./text-field/text-field.definition";
