+++ packages/web-components/fast-components/src/tab-panel/tab-panel.definition.ts
