+++ packages/web-components/fast-components/src/tab-panel/tab-panel.open-ui.definition.json
