+++ packages/web-components/fast-components/src/tab/tab.definition.ts
