+++ packages/web-components/fast-components/src/tab/tab.open-ui.definition.json
