+++ packages/web-components/fast-foundation/docs/api-report.md
@@ -481,11 +481,8 @@ export interface MenuItem extends StartEnd {
 
 // @public
 export enum MenuItemRole {
-    // (undocumented)
     menuitem = "menuitem",
-    // (undocumented)
     menuitemcheckbox = "menuitemcheckbox",
-    // (undocumented)
     menuitemradio = "menuitemradio"
 }
 
@@ -915,17 +912,11 @@ export class Tooltip extends FASTElement {
 
 // @public
 export enum TooltipPosition {
-    // (undocumented)
     bottom = "bottom",
-    // (undocumented)
     end = "end",
-    // (undocumented)
     left = "left",
-    // (undocumented)
     right = "right",
-    // (undocumented)
     start = "start",
-    // (undocumented)
     top = "top"
 }
