+++ packages/web-components/fast-foundation/src/divider/divider.options.ts
@@ -1,0 +1,15 @@
+/**
+ * Divider roles
+ * @public
+ */
+export enum DividerRole {
+    /**
+     * The divider semantically separates content
+     */
+    separator = "separator",
+
+    /**
+     * The divider has no semantic value and is for visual presentation only.
+     */
+    presentation = "presentation",
+}
