+++ packages/web-components/fast-foundation/src/divider/divider.spec.ts
@@ -1,14 +1,14 @@
 import { expect } from "chai";
-import { Divider, DividerTemplate as template } from "./index";
+import { customElement, DOM } from "@microsoft/fast-element";
 import { fixture } from "../fixture";
-import { DOM, customElement } from "@microsoft/fast-element";
-import { DividerRole } from "./divider";
+import { DividerRole } from "./divider.options";
+import { Divider, DividerTemplate as template } from "./index";
 
 @customElement({
     name: "fast-divider",
     template,
 })
-class FASTDivider extends Divider {}
+export class FASTDivider extends Divider {}
 
 async function setup() {
     const { element, connect, disconnect } = await fixture<Divider>("fast-divider");
