+++ packages/web-components/fast-foundation/src/flipper/flipper.options.ts
@@ -1,0 +1,8 @@
+/**
+ * The direction options for flipper.
+ * @public
+ */
+export enum FlipperDirection {
+    next = "next",
+    previous = "previous",
+}
