+++ packages/web-components/fast-foundation/src/flipper/flipper.spec.ts
@@ -1,14 +1,14 @@
 import { expect } from "chai";
-import { Flipper, FlipperTemplate as template } from "./index";
+import { customElement, DOM, html } from "@microsoft/fast-element";
 import { fixture } from "../fixture";
-import { DOM, customElement, html } from "@microsoft/fast-element";
-import { FlipperDirection } from "./flipper";
+import { FlipperDirection } from "./flipper.options";
+import { Flipper, FlipperTemplate as template } from "./index";
 
 @customElement({
     name: "fast-flipper",
     template,
 })
-class FASTFlipper extends Flipper {}
+export class FASTFlipper extends Flipper {}
 
 async function setup() {
     const { element, connect, disconnect } = await fixture<Flipper>("fast-flipper");
