+++ packages/web-components/fast-foundation/src/menu-item/menu-item.options.ts
@@ -1,0 +1,20 @@
+/**
+ * Menu items roles.
+ * @public
+ */
+export enum MenuItemRole {
+    /**
+     * The menu item has a "menuitem" role
+     */
+    menuitem = "menuitem",
+
+    /**
+     * The menu item has a "menuitemcheckbox" role
+     */
+    menuitemcheckbox = "menuitemcheckbox",
+
+    /**
+     * The menu item has a "menuitemradio" role
+     */
+    menuitemradio = "menuitemradio",
+}
