+++ packages/web-components/fast-foundation/src/text-area/text-area.options.ts
@@ -1,0 +1,25 @@
+/**
+ * Resize mode for a TextArea
+ * @public
+ */
+export enum TextAreaResize {
+    /**
+     * No resize.
+     */
+    none = "none",
+
+    /**
+     * Resize vertically and horizontally.
+     */
+    both = "both",
+
+    /**
+     * Resize horizontally.
+     */
+    horizontal = "horizontal",
+
+    /**
+     * Resize vertically.
+     */
+    vertical = "vertical",
+}
