+++ packages/web-components/fast-foundation/src/text-field/text-field.options.ts
@@ -1,0 +1,30 @@
+/**
+ * Text field sub-types
+ * @public
+ */
+export enum TextFieldType {
+    /**
+     * An email TextField
+     */
+    email = "email",
+
+    /**
+     * A password TextField
+     */
+    password = "password",
+
+    /**
+     * A telephone TextField
+     */
+    tel = "tel",
+
+    /**
+     * A text TextField
+     */
+    text = "text",
+
+    /**
+     * A URL TextField
+     */
+    url = "url",
+}
