+++ packages/web-components/fast-foundation/src/text-field/text-field.ts
@@ -2,37 +2,9 @@ import { attr, DOM, nullableNumberConverter, observable } from "@microsoft/fast-
 import { FormAssociated } from "../form-associated/form-associated";
 import { ARIAGlobalStatesAndProperties, StartEnd } from "../patterns/index";
 import { applyMixins } from "../utilities/index";
+import { TextFieldType } from "./text-field.options";
 
-/**
- * Text field sub-types
- * @public
- */
-export enum TextFieldType {
-    /**
-     * An email TextField
-     */
-    email = "email",
-
-    /**
-     * A password TextField
-     */
-    password = "password",
-
-    /**
-     * A telephone TextField
-     */
-    tel = "tel",
-
-    /**
-     * A text TextField
-     */
-    text = "text",
-
-    /**
-     * A URL TextField
-     */
-    url = "url",
-}
+export { TextFieldType };
 
 /**
  * An Text Field Custom HTML Element.
