+++ packages/web-components/fast-foundation/src/tooltip/tooltip.options.ts
@@ -1,0 +1,36 @@
+/**
+ * Enumerates possible tooltip positions
+ *
+ * @public
+ */
+export enum TooltipPosition {
+    /**
+     * The tooltip is positioned above the element
+     */
+    top = "top",
+
+    /**
+     * The tooltip is positioned to the right of the element
+     */
+    right = "right",
+
+    /**
+     * The tooltip is positioned below the element
+     */
+    bottom = "bottom",
+
+    /**
+     * The tooltip is positioned to the left of the element
+     */
+    left = "left",
+
+    /**
+     * The tooltip is positioned before the element
+     */
+    start = "start",
+
+    /**
+     * The tooltip is positioned after the element
+     */
+    end = "end",
+}
